backend/requirements.txt
flask
flask-cors
python-dotenv
groq
openai
gunicorn
